[fix] 모든 주식/코인 정수형으로 수정

See merge request s11-fintech-finance-sub1/S11P21A108!127

+++ backend/src/main/java/com/example/donzoom/dto/stock/response/StockTransactionHistoryResponseDto.java
@@ -12,10 +12,10 @@ public class StockTransactionHistoryResponseDto {
   private Long transactionHistoryId;  // 거래 내역 ID
   private Long stockId;               // 주식 ID
   private TransactionType transactionType;  // 거래 타입
-  private Float price;                // 매수 또는 매도 가격
+  private Integer price;                // 매수 또는 매도 가격
   private Integer amount;             // 매수 또는 매도 수량
-  private Float total;                // 총 거래 금액 (price * amount)
-  private Float profit;       // 실현 수익 (매도 시에만 기록)
+  private Integer total;                // 총 거래 금액 (price * amount)
+  private Integer profit;       // 실현 수익 (매도 시에만 기록)
   private LocalDateTime createdAt;    // 거래 생성 시간
 
 }
